Migrate Mutation.updateOrganizationName
